requirements.txt
dj-database-url==0.5.0
Django==2.0.5
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.0
django-crispy-forms==1.7.2
djangorestframework
requests
watson-developer-cloud
braintree
social-auth-app-django==2.1.0
django-cors-headers==2.1.0
django-filter==2.0.0
djangorestframework-jwt==1.11.0
pyJWT==1.5.3